better multitask for kernel(untested)

+++ kernel/src/filesystem/dev/null.rs
@@ -11,7 +11,7 @@ impl InodeOperation for NullDevice {
         buf.len()
     }
 
-    fn len(&self) -> usize {
+    fn len(&self) -> u64 {
         0
     }
 }
