+++ kernel/src/filesystem/dev/zero.rs
@@ -12,7 +12,7 @@ impl InodeOperation for ZeroDevice {
         buf.len()
     }
 
-    fn len(&self) -> usize {
+    fn len(&self) -> u64 {
         0
     }
 }
